requirements.txt
# requirements.txt
fastapi==0.128.6
uvicorn[standard]==0.40.0
mcp==1.26.0
boto3==1.42.45
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1
jinja2==3.1.6
rich==14.3.2
GitPython==3.1.46